refactor: rearrange classes
