+++ pydid/__init__.py
@@ -3,11 +3,15 @@
 from .common import DIDError
 from .did import DID, InvalidDIDError
 from .did_url import DIDUrl, InvalidDIDUrlError
-from .doc import DIDDocumentError
 from .doc.doc import DIDDocument, BasicDIDDocument
 from .doc.builder import DIDDocumentBuilder
-from .doc.verification_method import VerificationMethod
-from .doc.service import Service, DIDCommService
+from .verification_method import (
+    VerificationMethod,
+    VerificationMaterial,
+    VerificationMaterialUnknown,
+)
+from .service import Service, DIDCommService
+from .resource import Resource
 
 __all__ = [
     "BasicDIDDocument",
@@ -22,4 +26,7 @@ __all__ = [
     "InvalidDIDUrlError",
     "Service",
     "VerificationMethod",
+    "VerificationMaterial",
+    "VerificationMaterialUnknown",
+    "Resource",
 ]
