+++ pydid/doc/__init__.py
@@ -1,5 +1,30 @@
 """DID Document classes."""
 
+from .doc import (
+    IdentifiedResourceMismatch,
+    IDNotFoundError,
+    DIDDocumentRoot,
+    BasicDIDDocument,
+    DIDDocument,
+)
 
-class DIDDocumentError(Exception):
-    """Base Exception for problems with DID Documents."""
+from .builder import (
+    VerificationMethodBuilder,
+    RelationshipBuilder,
+    ServiceBuilder,
+    DIDDocumentBuilder,
+)
+
+
+__all__ = [
+    "DIDDocumentError",
+    "IdentifiedResourceMismatch",
+    "IDNotFoundError",
+    "DIDDocumentRoot",
+    "BasicDIDDocument",
+    "DIDDocument",
+    "VerificationMethodBuilder",
+    "RelationshipBuilder",
+    "ServiceBuilder",
+    "DIDDocumentBuilder",
+]
