--- tests/doc/test_resource.py
@@ -1,5 +1,0 @@
-from pydid.doc.resource import Resource
-
-
-def test_resource():
-    Resource()
