+++ tests/test_resource.py
@@ -1,0 +1,5 @@
+from pydid.resource import Resource
+
+
+def test_resource():
+    Resource()
